+++ src/gemseo/problems/sobieski/wrappers_sg.py
@@ -75,10 +75,10 @@ class SobieskiMissionSG(SobieskiBaseWrapperSimpleGram):
         :type dtype: str
         """
         super(SobieskiMissionSG, self).__init__(dtype=dtype)
-        self.input_grammar.add_elements(
+        self.input_grammar.update_elements(
             **dict.fromkeys(("y_14", "y_24", "y_34", "x_shared"), ndarray)
         )
-        self.output_grammar.add_elements(y_4=ndarray)
+        self.output_grammar.update_elements(y_4=ndarray)
         self._set_default_inputs()
 
     def _run(self):
@@ -108,10 +108,10 @@ class SobieskiStructureSG(SobieskiBaseWrapperSimpleGram):
     def __init__(self, dtype=DTYPE_DOUBLE):
         """Constructor of wrapper for weight computation."""
         super(SobieskiStructureSG, self).__init__(dtype=dtype)
-        self.input_grammar.add_elements(
+        self.input_grammar.update_elements(
             **dict.fromkeys(["x_1", "y_21", "y_31", "x_shared"], ndarray)
         )
-        self.output_grammar.add_elements(
+        self.output_grammar.update_elements(
             **dict.fromkeys(["y_1", "y_11", "y_12", "y_14", "g_1"], ndarray)
         )
         self._set_default_inputs()
@@ -150,10 +150,10 @@ class SobieskiAerodynamicsSG(SobieskiBaseWrapperSimpleGram):
         :type dtype: str
         """
         super(SobieskiAerodynamicsSG, self).__init__(dtype=dtype)
-        self.input_grammar.add_elements(
+        self.input_grammar.update_elements(
             **dict.fromkeys(["x_2", "y_12", "y_32", "x_shared"], ndarray)
         )
-        self.output_grammar.add_elements(
+        self.output_grammar.update_elements(
             **dict.fromkeys(["y_2", "y_21", "y_23", "y_24", "g_2"], ndarray)
         )
         self._set_default_inputs()
@@ -191,10 +191,10 @@ class SobieskiPropulsionSG(SobieskiBaseWrapperSimpleGram):
         :type dtype: str
         """
         super(SobieskiPropulsionSG, self).__init__(dtype=dtype)
-        self.input_grammar.add_elements(
+        self.input_grammar.update_elements(
             **dict.fromkeys(["x_3", "y_23", "x_shared"], ndarray)
         )
-        self.output_grammar.add_elements(
+        self.output_grammar.update_elements(
             **dict.fromkeys(["y_3", "y_34", "y_31", "y_32", "g_3"], ndarray)
         )
         self._set_default_inputs()